# docs/references/api/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is the API specification of the Epinio API",
    "title": "Epinio API",
    "version": "1.0.0"
  },
  "basePath": "/api/v1",
  "paths": {
...
    "/namespaces/{Namespace}/applications": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Return list of applications in the `Namespace`.",
        "operationId": "Apps",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppsResponse"
          }
        }
      },
      "post": {
        "tags": [
          "application"
        ],
        "summary": "Create the posted new application in the `Namespace`.",
        "operationId": "AppCreate",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "name": "Configuration",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ApplicationCreateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppCreateResponse"
          }
        }
      },
      "delete": {
        "summary": "Delete the named `Applications` in the `Namespace`.",
        "operationId": "AppBatchDelete",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "in: url",
            "name": "Applications",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppDeleteResponse"
          }
        }
      }
    },
    "/namespaces/{Namespace}/applications/{App}": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Return details of the named `App` in the `Namespace`.",
        "operationId": "AppShow",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppShowResponse"
          }
        }
      },
      "delete": {
        "tags": [
          "application"
        ],
        "summary": "Delete the named `App` in the `Namespace`.",
        "operationId": "AppDelete",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppDeleteResponse"
          }
        }
      },
      "patch": {
        "tags": [
          "application"
        ],
        "summary": "Patch the named `App` in the `Namespace`.",
        "operationId": "AppUpdate",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ApplicationUpdateRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppUpdateResponse"
          }
        }
      }
    },
...
    "/namespaces/{Namespace}/applications/{App}/logs": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Return logs of the named `App` in the `Namespace` streamed over a websocket.",
        "operationId": "AppLogs",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "App",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AppLogsResponse"
          }
        }
      }
    },
...
    "/namespaces/{Namespace}/staging/{StageID}/logs": {
      "get": {
        "tags": [
          "application"
        ],
        "summary": "Return logs of the named `StageID` in the `Namespace` streamed over a websocket.",
        "operationId": "StagingLogs",
        "parameters": [
          {
            "type": "string",
            "name": "Namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "StageID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/StagingLogsResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "App": {
      "description": "The main structure has identifying information.\nIt is used in the CLI and API responses.",
      "type": "object",
      "title": "App has all the application's properties, for at rest (Configuration), and active (Workload).",
      "properties": {
        "configuration": {
          "$ref": "#/definitions/ApplicationConfiguration"
        },
        "deployment": {
          "$ref": "#/definitions/AppDeployment"
        },
        "image_url": {
          "type": "string",
          "x-go-name": "ImageURL"
        },
        "meta": {
          "$ref": "#/definitions/AppRef"
        },
        "origin": {
          "$ref": "#/definitions/ApplicationOrigin"
        },
        "stage_id": {
          "type": "string",
          "x-go-name": "StageID"
        },
        "staging": {
          "$ref": "#/definitions/ApplicationStage"
        },
        "stagingstatus": {
          "$ref": "#/definitions/ApplicationStagingStatus"
        },
        "status": {
          "$ref": "#/definitions/ApplicationStatus"
        },
        "statusmessage": {
          "type": "string",
          "x-go-name": "StatusMessage"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "AppDeployment": {
      "description": "AppDeployment contains all the information specific to an active\napplication, i.e. one with a deployment in the cluster.",
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean",
          "x-go-name": "Active"
        },
        "createdAt": {
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "desiredreplicas": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "DesiredReplicas"
        },
        "name": {
          "description": "TODO: Readiness and Liveness fields?",
          "type": "string",
          "x-go-name": "Name"
        },
        "readyreplicas": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "ReadyReplicas"
        },
        "replicas": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/PodInfo"
          },
          "x-go-name": "Replicas"
        },
        "routes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Routes"
        },
        "stage_id": {
          "type": "string",
          "x-go-name": "StageID"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "AppList": {
      "description": "AppList is a collection of app references",
      "type": "array",
      "items": {
        "$ref": "#/definitions/App"
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "AppRef": {
      "description": "AppRef references an App by name and namespace",
      "type": "object",
      "properties": {
        "createdAt": {
          "$ref": "#/definitions/Time"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "namespace": {
          "type": "string",
          "x-go-name": "Namespace"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationConfiguration": {
      "description": "ApplicationConfiguration is the part of the manifest describing the configuration of the application",
      "type": "object",
      "properties": {
        "appchart": {
          "type": "string",
          "x-go-name": "AppChart"
        },
        "configurations": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Configurations"
        },
        "environment": {
          "$ref": "#/definitions/EnvVariableMap"
        },
        "instances": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "Instances"
        },
        "routes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Routes"
        },
        "settings": {
          "$ref": "#/definitions/ChartValueSettings"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationCreateRequest": {
      "description": "ApplicationCreateRequest represents and contains the data needed to\ncreate an application (at rest), possibly with presets (configurations)",
      "type": "object",
      "properties": {
        "configuration": {
          "$ref": "#/definitions/ApplicationUpdateRequest"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationDeleteResponse": {
      "description": "ApplicationDeleteResponse represents the server's response to a successful app deletion",
      "type": "object",
      "properties": {
        "unboundconfigurations": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "UnboundConfigurations"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationOrigin": {
      "description": "ApplicationOrigin is the part of the manifest describing the origin of the application\n(sources). At most one of the fields may be specified / not empty.",
      "type": "object",
      "properties": {
        "Kind": {
          "description": "Hidden from yaml. Type tag to simplify struct usage.\nNote: we cannot hide this property from the JSON since it's used to unmarshal correctly the result of the Apps endpoint\n@see failling test here: https://github.com/epinio/epinio/runs/4935898437?check_suite_focus=true\nWe should probably expose a more meaningful value instead of this \"Kind\" int",
          "type": "integer",
          "format": "int64"
        },
        "archive": {
          "type": "boolean",
          "x-go-name": "Archive"
        },
        "container": {
          "type": "string",
          "x-go-name": "Container"
        },
        "git": {
          "$ref": "#/definitions/GitRef"
        },
        "path": {
          "type": "string",
          "x-go-name": "Path"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationStage": {
      "description": "ApplicationStage is the part of the manifest holding information\nrelevant to staging the application's sources. This is, currently,\nonly the reference to the Paketo builder image to use.",
      "type": "object",
      "properties": {
        "builder": {
          "type": "string",
          "x-go-name": "Builder"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationStagingStatus": {
      "type": "string",
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationStatus": {
      "type": "string",
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "ApplicationUpdateRequest": {
      "description": "ApplicationUpdateRequest represents and contains the data needed to update\nan application. Specifically to modify the number of replicas to\nrun, and the configurations bound to it.\nNote: Instances is a pointer to give us a nil value separate from\nactual integers, as means of communicating `default`/`no change`.",
      "type": "object",
      "properties": {
        "appchart": {
          "type": "string",
          "x-go-name": "AppChart"
        },
        "configurations": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Configurations"
        },
        "environment": {
          "$ref": "#/definitions/EnvVariableMap"
        },
        "instances": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "Instances"
        },
        "restart": {
          "type": "boolean",
          "x-go-name": "Restart"
        },
        "routes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Routes"
        },
        "settings": {
          "$ref": "#/definitions/ChartValueSettings"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "ChartValueSettings": {
      "description": "ChartValueSettings is a collection of key/value pairs describing the user's chosen settings with which\nto configure the helm chart referenced by the application's appchart.",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "EnvVariableMap": {
      "description": "EnvVariableMap is a collection of EVs as a map. It is used for Set Requests, and as\nList Responses",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "PodInfo": {
      "type": "object",
      "properties": {
        "createdAt": {
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "memoryBytes": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "MemoryBytes"
        },
        "metricsOk": {
          "type": "boolean",
          "x-go-name": "MetricsOk"
        },
        "millicpus": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "MilliCPUs"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "ready": {
          "type": "boolean",
          "x-go-name": "Ready"
        },
        "restarts": {
          "type": "integer",
          "format": "int32",
          "x-go-name": "Restarts"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
    "Response": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-package": "github.com/epinio/epinio/pkg/api/core/v1/models"
    },
...
    "Time": {
      "description": "+protobuf.options.marshal=false\n+protobuf.as=Timestamp\n+protobuf.options.(gogoproto.goproto_stringer)=false",
      "type": "object",
      "title": "Time is a wrapper around time.Time which supports correct\nmarshaling to YAML and JSON.  Wrappers are provided for many\nof the factory methods that the time package offers.",
      "x-go-package": "k8s.io/apimachinery/pkg/apis/meta/v1"
    },
...
  },
  "responses": {
...
    "AppCreateResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
    "AppDeleteResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/ApplicationDeleteResponse"
      }
    },
...
    "AppLogsResponse": {
      "description": ""
    },
...
    "AppShowResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/App"
      }
    },
...
    "AppUpdateResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Response"
      }
    },
...
    "AppsResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/AppList"
      }
    },
...
    "StagingLogsResponse": {
      "description": ""
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basicAuth": [
        "[]"
      ]
    }
  ]
}